GetProductListRes :  fixed crashes

+++ app/src/main/java/com/app/ulife/creator/models/productList/GetProductListRes.kt
@@ -15,7 +15,7 @@ data class GetProductListRes(
 )
 
 data class Data(
-    val bv: Int,
+    val bv: Double,
     val desc: String,
     val img: String,
     val mrp: Double,
